Introduce crypto::sign, get rid of crypto::misc.
