+++ src/crypto/common.rs
@@ -8,7 +8,6 @@ use core::fmt;
 use std::error;
 use std::vec::Vec;
 
-use crate::base::iana::SecAlg;
 use crate::rdata::Dnskey;
 
 #[cfg(feature = "openssl")]
@@ -16,54 +15,6 @@ use super::openssl;
 
 #[cfg(feature = "ring")]
 use super::ring;
-
-//----------- GenerateParams -------------------------------------------------
-
-/// Parameters for generating a secret key.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum GenerateParams {
-    /// Generate an RSA/SHA-256 keypair.
-    RsaSha256 {
-        /// The number of bits in the public modulus.
-        ///
-        /// A ~3000-bit key corresponds to a 128-bit security level.  However,
-        /// RSA is mostly used with 2048-bit keys.  Some backends (like Ring)
-        /// do not support smaller key sizes than that.
-        ///
-        /// For more information about security levels, see [NIST SP 800-57
-        /// part 1 revision 5], page 54, table 2.
-        ///
-        /// [NIST SP 800-57 part 1 revision 5]: https://nvlpubs.nist.gov/nistpubs/SpecialPublications/NIST.SP.800-57pt1r5.pdf
-        bits: u32,
-    },
-
-    /// Generate an ECDSA P-256/SHA-256 keypair.
-    EcdsaP256Sha256,
-
-    /// Generate an ECDSA P-384/SHA-384 keypair.
-    EcdsaP384Sha384,
-
-    /// Generate an Ed25519 keypair.
-    Ed25519,
-
-    /// An Ed448 keypair.
-    Ed448,
-}
-
-//--- Inspection
-
-impl GenerateParams {
-    /// The algorithm of the generated key.
-    pub fn algorithm(&self) -> SecAlg {
-        match self {
-            Self::RsaSha256 { .. } => SecAlg::RSASHA256,
-            Self::EcdsaP256Sha256 => SecAlg::ECDSAP256SHA256,
-            Self::EcdsaP384Sha384 => SecAlg::ECDSAP384SHA384,
-            Self::Ed25519 => SecAlg::ED25519,
-            Self::Ed448 => SecAlg::ED448,
-        }
-    }
-}
 
 //----------- DigestType -----------------------------------------------------
 
@@ -290,151 +241,3 @@ impl fmt::Display for ParseDnskeyTextError {
 }
 
 impl error::Error for ParseDnskeyTextError {}
-
-#[cfg(feature = "unstable-crypto-sign")]
-pub mod sign {
-    use std::vec::Vec;
-
-    use crate::base::iana::SecAlg;
-    use crate::crypto::misc::{
-        FromBytesError, GenerateError, SecretKeyBytes, SignError, SignRaw,
-        Signature,
-    };
-    use crate::rdata::Dnskey;
-
-    use super::GenerateParams;
-
-    #[cfg(feature = "openssl")]
-    use super::openssl;
-
-    #[cfg(feature = "ring")]
-    use super::ring;
-
-    //----------- KeyPair ----------------------------------------------------
-
-    /// A key pair based on a built-in backend.
-    ///
-    /// This supports any built-in backend (currently, that is OpenSSL and Ring,
-    /// if their respective feature flags are enabled).  Wherever possible, it
-    /// will prefer the Ring backend over OpenSSL -- but for more uncommon or
-    /// insecure algorithms, that Ring does not support, OpenSSL must be used.
-    #[derive(Debug)]
-    // Note: ring does not implement Clone for KeyPair.
-    pub enum KeyPair {
-        /// A key backed by Ring.
-        #[cfg(feature = "ring")]
-        Ring(ring::sign::KeyPair),
-
-        /// A key backed by OpenSSL.
-        #[cfg(feature = "openssl")]
-        OpenSSL(openssl::sign::KeyPair),
-    }
-
-    //--- Conversion to and from bytes
-
-    impl KeyPair {
-        /// Import a key pair from bytes.
-        pub fn from_bytes<Octs>(
-            secret: &SecretKeyBytes,
-            public: &Dnskey<Octs>,
-        ) -> Result<Self, FromBytesError>
-        where
-            Octs: AsRef<[u8]>,
-        {
-            // Prefer Ring if it is available.
-            #[cfg(feature = "ring")]
-            {
-                let fallback_to_openssl = match public.algorithm() {
-                    SecAlg::RSASHA1
-                    | SecAlg::RSASHA1_NSEC3_SHA1
-                    | SecAlg::RSASHA256
-                    | SecAlg::RSASHA512 => {
-                        ring::PublicKey::from_dnskey(public)
-                            .map_err(|_| FromBytesError::InvalidKey)?
-                            .key_size()
-                            < 2048
-                    }
-                    _ => false,
-                };
-
-                if !fallback_to_openssl {
-                    let key =
-                        ring::sign::KeyPair::from_bytes(secret, public)?;
-                    return Ok(Self::Ring(key));
-                }
-            }
-
-            // Fall back to OpenSSL.
-            #[cfg(feature = "openssl")]
-            return Ok(Self::OpenSSL(openssl::sign::KeyPair::from_bytes(
-                secret, public,
-            )?));
-
-            // Otherwise fail.
-            #[allow(unreachable_code)]
-            Err(FromBytesError::UnsupportedAlgorithm)
-        }
-    }
-
-    //--- SignRaw
-
-    impl SignRaw for KeyPair {
-        fn algorithm(&self) -> SecAlg {
-            match self {
-                #[cfg(feature = "ring")]
-                Self::Ring(key) => key.algorithm(),
-                #[cfg(feature = "openssl")]
-                Self::OpenSSL(key) => key.algorithm(),
-            }
-        }
-
-        fn dnskey(&self) -> Dnskey<Vec<u8>> {
-            match self {
-                #[cfg(feature = "ring")]
-                Self::Ring(key) => key.dnskey(),
-                #[cfg(feature = "openssl")]
-                Self::OpenSSL(key) => key.dnskey(),
-            }
-        }
-
-        fn sign_raw(&self, data: &[u8]) -> Result<Signature, SignError> {
-            match self {
-                #[cfg(feature = "ring")]
-                Self::Ring(key) => key.sign_raw(data),
-                #[cfg(feature = "openssl")]
-                Self::OpenSSL(key) => key.sign_raw(data),
-            }
-        }
-    }
-
-    //----------- generate() -------------------------------------------------
-
-    /// Generate a new secret key for the given algorithm.
-    pub fn generate(
-        params: GenerateParams,
-        flags: u16,
-    ) -> Result<(SecretKeyBytes, Dnskey<Vec<u8>>), GenerateError> {
-        // Use Ring if it is available.
-        #[cfg(feature = "ring")]
-        if matches!(
-            &params,
-            GenerateParams::EcdsaP256Sha256
-                | GenerateParams::EcdsaP384Sha384
-                | GenerateParams::Ed25519
-        ) {
-            let rng = ::ring::rand::SystemRandom::new();
-            return Ok(ring::sign::generate(params, flags, &rng)?);
-        }
-
-        // Fall back to OpenSSL.
-        #[cfg(feature = "openssl")]
-        {
-            let key = openssl::sign::generate(params, flags)?;
-            return Ok((key.to_bytes(), key.dnskey()));
-        }
-
-        // Otherwise fail.
-        #[allow(unreachable_code)]
-        Err(GenerateError::UnsupportedAlgorithm)
-    }
-}
